Sensors: Fix an improper type definition when loading compass gain.
To load compass gain with a correct type, we should fix it to 'int'.
This should be common fix but partial related to bug 7355959.

Bug: 7355959

+++ drivers/staging/iio/magnetometer/inv_ami306_ring.c
@@ -50,7 +50,7 @@ enum Compass_cali_File {
 };
 
 /* function for loading compass calibration file. */
-static int access_cali_file(short *gain, int target)
+static int access_cali_file(int *gain, int target)
 {
 	char buf[256];
 	int ret;
